add address field to UnspentOutput stuct
