+++ app/walletmediums/dcrwalletrpc/walletfunctions.go
@@ -148,6 +148,11 @@ func (c *WalletPRCClient) UnspentOutputs(account uint32, targetAmount int64) ([]
 		}
 		txHash := hash.String()
 
+		address, err := walletcore.GetAddressFromPkScript(utxo.PkScript)
+		if err != nil {
+			return nil, err
+		}
+
 		unspentOutput := &walletcore.UnspentOutput{
 			OutputKey:       fmt.Sprintf("%s:%d", txHash, utxo.OutputIndex),
 			TransactionHash: txHash,
@@ -155,7 +160,7 @@ func (c *WalletPRCClient) UnspentOutputs(account uint32, targetAmount int64) ([]
 			Tree:            utxo.Tree,
 			ReceiveTime:     utxo.ReceiveTime,
 			Amount:          dcrutil.Amount(utxo.Amount),
-			PkScript:        utxo.PkScript,
+			Address:         address,
 		}
 		unspentOutputs = append(unspentOutputs, unspentOutput)
 	}
